+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - `ext:install` now supports `--force` and `--non-interactive` flags.
 - Fixes a crash when customers deploy an empty Cloud Functions project (#3705)
+- Fixes (and implements) `--no-authorized-domains`, skipping syncing with Firebase Auth, when deploying to a Firebase Hosting channel (#3740).
 - Adds a command (`functions:list`) for listing all functions in the Firebase project.